add MG-RAST test, MG-RAST sample listing (close #13)

+++ .circleci/test.sh
@@ -13,17 +13,24 @@ python bin/grabseqs -h
 # repo grabseqs tests
 mkdir -p $HOME/tmp_test
 
+# SRA
 ## test sample listing, metadata download
 if [ `python bin/grabseqs sra -m -l -o $HOME/tmp_test/test_metadata/ SRP057027 | wc -l` -ne 369 ]; then
     exit 1
 fi
-echo ":D  sample listing test passed"
-
+echo ":D SRA sample listing test passed"
 ## test metadata download
 if [ `cat $HOME/tmp_test/test_metadata/SRP057027.tsv | wc -l` -ne 371 ] ; then
     exit 1
 fi
-echo ":D  metadata test passed"
+echo ":D SRA metadata test passed"
+
+# MG-RAST
+## test sample listing
+if [ `python bin/grabseqs mgrast -l mgp8384 | wc -l` -ne 12 ]; then
+    exit 1
+fi
+echo ":D MG-RAST sample listing test passed"
 
 # test conda install
 conda install -c louiejtaylor grabseqs > /dev/null
